services/spend-guard/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
google-cloud-bigquery==3.25.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
